+++ src/User/UserService.php
@@ -40,7 +40,7 @@ final class UserService extends AbstractService
      *
      * @throws AuthenticationRequired
      */
-    public function updateUser(int $userId, string $email, string $firstName, string $lastName)
+    public function updateUser(int $userId, string $email, string $firstName, string $lastName, ?UserTitle $title = null)
     {
         $this->client->mustBeAuthenticated();
         $this->client->put(
@@ -48,6 +48,7 @@ final class UserService extends AbstractService
             [
                 'form_params' => [
                     'email' => $email,
+                    'title' => is_null($title) ? null : $title->getValue(),
                     'firstName' => $firstName,
                     'lastName' => $lastName,
                 ],